+++ demo/CMakeFiles/Makefile.cmake
@@ -155,12 +155,13 @@ set(CMAKE_MAKEFILE_PRODUCTS
 
 # Dependency information for all targets:
 set(CMAKE_DEPEND_INFO_FILES
-  "CMakeFiles/testRenderMRI.dir/DependInfo.cmake"
-  "CMakeFiles/testRenderAndImage.dir/DependInfo.cmake"
-  "CMakeFiles/testMRIRender.dir/DependInfo.cmake"
-  "CMakeFiles/testCutter.dir/DependInfo.cmake"
-  "CMakeFiles/testContour.dir/DependInfo.cmake"
-  "CMakeFiles/FixedPointVolumeRayCastMapperCT.dir/DependInfo.cmake"
   "CMakeFiles/testMapper.dir/DependInfo.cmake"
   "CMakeFiles/volumeRender.dir/DependInfo.cmake"
+  "CMakeFiles/testContour.dir/DependInfo.cmake"
+  "CMakeFiles/testCutter.dir/DependInfo.cmake"
+  "CMakeFiles/testMRIRender.dir/DependInfo.cmake"
+  "CMakeFiles/testRenderAndImage.dir/DependInfo.cmake"
+  "CMakeFiles/testRenderMRI.dir/DependInfo.cmake"
+  "CMakeFiles/MRIRender.dir/DependInfo.cmake"
+  "CMakeFiles/FixedPointVolumeRayCastMapperCT.dir/DependInfo.cmake"
   )
