+++ demo/MRIRender.cxx
@@ -20,160 +20,9 @@
 #include <vtkRenderWindowInteractor.h>
 #include <vtkInteractorObserver.h>
 #include <vtkRenderer.h>
-#include <vtkInteractorStyleImage.h>
-#include <vtkObjectFactory.h>
 
 
 #include "utility_RenderAndImage.h"
-
-
-
-
-class myVtkInteractorStyleImage : public vtkInteractorStyleImage
-{
-public:
-  static myVtkInteractorStyleImage* New();
-  vtkTypeMacro(myVtkInteractorStyleImage, vtkInteractorStyleImage);
-  vtkImageData* _MRI_Copy;
-  vtkImageData* _oneFrame;
-  vtkImageData* _MRI_image;
-  vtkVolume* _MRI_ROI_volume;
-  vtkImageActor* _imageActor;
-  int _MinSlice;
-  int _MaxSlice;
- 
-protected:
-  vtkRenderer* _renderer;
-  int _Slice;
-  int* dims; 
- 
-public:
-   void SetRender(vtkRenderer* renderer) {
-     _renderer = renderer;
-     _Slice = _MinSlice;
-     ShowFrame(_Slice);
-     dims = _MRI_Copy->GetDimensions();
-     _renderer->AddActor(_imageActor);
-     _renderer->AddVolume(_MRI_ROI_volume);
-   }
- 
- 
-protected:
-
-   void ShowFrame(int FrameNo){
-     for(int i = 0; i < dims[0]; i++)
-       for (int j = 0; j < dims[1]; j++)
-         for (int k = 0; k < dims[2]; k++)
-           {
-             float* oneFrame_pixel =
-               static_cast<float*>(_oneFrame->GetScalarPointer(i, j, 0));
-             float* MRI_Copypixel =
-               static_cast<float*>(_MRI_Copy->GetScalarPointer(i, j, k));
-             if (k == FrameNo)
-               {
-                 float* MRI_pixel =
-                   static_cast<float*>(_MRI_image->GetScalarPointer(i, j, k));
-                 oneFrame_pixel[0] = MRI_pixel[0];
-               }
-             if (k < FrameNo)
-               MRI_Copypixel[0] = 0.0;
-           }
-     vtkSmartPointer<vtkFixedPointVolumeRayCastMapper> MRI_ROI_mapper =
-       vtkSmartPointer<vtkFixedPointVolumeRayCastMapper>::New();
-     MRI_ROI_mapper->SetInputData(_MRI_Copy);
-     MRI_ROI_mapper->SetBlendModeToMaximumIntensity();
-  
-
-     vtkSmartPointer<vtkVolumeProperty> MRI_ROI_Property =
-       property(0, 255, 1, 'B');
-
-     
-     _MRI_ROI_volume->SetMapper(MRI_ROI_mapper);
-     _MRI_ROI_volume->SetProperty(MRI_ROI_Property);
-
-  
-
-     vtkSmartPointer<vtkImageResliceMapper> imageMapper =
-       vtkSmartPointer<vtkImageResliceMapper>::New();
-     imageMapper->SetInputData(_oneFrame);
-
-    
-     _imageActor->GetMapper()->SetInputData(_oneFrame);
-     _imageActor->SetPosition(0, 0, FrameNo*2);
-     _imageActor->SetInterpolate(1);
-   }
-   /*void ShowGradient(){
-     _ImageViewer->SetInputConnection(_gradientFilter->GetOutputPort());
-     _ImageViewer->SetSlice(_Slice);
-     _ImageViewer->Render();
-   }
-   void ShowOrigin(){
-    _ImageViewer->SetInputConnection(_shiftScaleFilter->GetOutputPort());
-     _ImageViewer->SetSlice(_Slice);
-     _ImageViewer->Render();
-     }*/
-   void MoveSliceForward() {
-      if(_Slice < _MaxSlice) {
-         _Slice += 1;
-         ShowFrame(_Slice);
-         _renderer->AddActor(_imageActor);
-         _renderer->AddVolume(_MRI_ROI_volume);
-      }
-   }
- 
-   void MoveSliceBackward() {
-      if(_Slice > _MinSlice) {
-         _Slice -= 1;
-         ShowFrame(_Slice);
-         _renderer->AddActor(_imageActor);
-         _renderer->AddVolume(_MRI_ROI_volume);
-      }
-   }
- 
- 
-   virtual void OnKeyDown() {
-      std::string key = this->GetInteractor()->GetKeySym();
-      if(key.compare("Up") == 0) {
-         //cout << "Up arrow key was pressed." << endl;
-         MoveSliceForward();
-      }
-      else if(key.compare("Down") == 0) {
-         //cout << "Down arrow key was pressed." << endl;
-         MoveSliceBackward();
-      }
-      /* else if(key.compare("g") == 0) {
-	ShowGradient();
-      }
-      else if(key.compare("o") == 0) {
-      ShowOrigin();
-      }*/
-      // forward event
-      vtkInteractorStyleImage::OnKeyDown();
-   }
- 
- 
-   virtual void OnMouseWheelForward() {
-      //std::cout << "Scrolled mouse wheel forward." << std::endl;
-      MoveSliceForward();
-      // don't forward events, otherwise the image will be zoomed 
-      // in case another interactorstyle is used (e.g. trackballstyle, ...)
-      // vtkInteractorStyleImage::OnMouseWheelForward();
-   }
- 
- 
-   virtual void OnMouseWheelBackward() {
-      //std::cout << "Scrolled mouse wheel backward." << std::endl;
-      if(_Slice > _MinSlice) {
-         MoveSliceBackward();
-      }
-      // don't forward events, otherwise the image will be zoomed 
-      // in case another interactorstyle is used (e.g. trackballstyle, ...)
-      // vtkInteractorStyleImage::OnMouseWheelBackward();
-   }
-};
-
-vtkStandardNewMacro(myVtkInteractorStyleImage);
-
 
 
 
@@ -332,14 +181,30 @@ int main (int argc, char *argv[])
 
   std::cout << Z_min << " " << Z_max << std::endl;
 
+  vtkSmartPointer<vtkFixedPointVolumeRayCastMapper> MRI_ROI_mapper =
+    vtkSmartPointer<vtkFixedPointVolumeRayCastMapper>::New();
+  MRI_ROI_mapper->SetInputData(MRI_Copy);
+  MRI_ROI_mapper->SetBlendModeToMaximumIntensity();
+  
+
+  vtkSmartPointer<vtkVolumeProperty> MRI_ROI_Property =
+    property(0, 255, independentComponents, 'B');
+
   vtkSmartPointer<vtkVolume> MRI_ROI_volume =
     vtkSmartPointer<vtkVolume>::New();
-  vtkSmartPointer<vtkImageActor> imageActor =
-    vtkSmartPointer<vtkImageActor>::New();
+  MRI_ROI_volume->SetMapper(MRI_ROI_mapper);
+  MRI_ROI_volume->SetProperty(MRI_ROI_Property);
 
-  // my interactorStyle
-  vtkSmartPointer<myVtkInteractorStyleImage> myInteractorStyle =
-    vtkSmartPointer<myVtkInteractorStyleImage>::New();
+  
+
+  vtkSmartPointer<vtkImageResliceMapper> imageMapper =
+    vtkSmartPointer<vtkImageResliceMapper>::New();
+  imageMapper->SetInputData(oneFrame);
+
+  vtkSmartPointer<vtkImageActor> imageActor = vtkSmartPointer<vtkImageActor>::New();
+  imageActor->GetMapper()->SetInputData(oneFrame);
+  imageActor->SetPosition(0, 0, Z_min*2);
+  imageActor->SetInterpolate(1);
 
   // Create the renderers, render window, and interactor
 
@@ -349,26 +214,16 @@ int main (int argc, char *argv[])
     vtkSmartPointer<vtkRenderWindowInteractor>::New();
   vtkSmartPointer<vtkRenderer> ren =
     vtkSmartPointer<vtkRenderer>::New();
-
-  myInteractorStyle->_imageActor = imageActor;
-  myInteractorStyle->SetRender(ren);
-  myInteractorStyle->_MRI_Copy = MRI_Copy;
-  myInteractorStyle->_oneFrame = oneFrame;
-  myInteractorStyle->_MRI_image = MRI_image;
-  myInteractorStyle->_MRI_ROI_volume = MRI_ROI_volume;
-  myInteractorStyle->_imageActor = imageActor;
-  myInteractorStyle->_MinSlice = Z_min;
-  myInteractorStyle->_MaxSlice = Z_max;
-
   renWin->AddRenderer(ren);
   iren->SetRenderWindow(renWin);
   iren->SetDesiredUpdateRate(frameRate / (1+clip));
   iren->GetInteractorStyle()->SetDefaultRenderer(ren);
-  iren->SetInteractorStyle(myInteractorStyle);
 
   
   renWin->SetSize(600, 600);
   
+  ren->AddActor(imageActor);
+  ren->AddVolume(MRI_ROI_volume);
   renWin->Render();
   iren->Start();
